backend/requirements.txt
fastapi[all]
pypdf2
sentence-transformers
torch
chromadb
python-multipart
firebase-admin
langchain
langchain-text-splitters